+++ src/track/trackwavpack.h
@@ -16,7 +16,7 @@ protected:
     void close() override;
     void seek(uint ms) override;
     uint currentPos() override;
-    int readSamples(soundtouch::SAMPLETYPE *buffer, unsigned int num) override;
+    int readSamples(soundtouch::SAMPLETYPE *buffer, size_t num) override;
 
     void storeBPM(std::string sBPM) override;
     void removeBPM() override;
